Backend/requirements.txt
alembic==1.13.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachelib==0.12.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.15.4
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
identify==2.6.0
idna==3.7
iniconfig==2.0.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
Mako==1.3.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgspec==0.18.6
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.9.1
numerize==0.12
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas<2.0.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prompt-toolkit==3.0.38
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.15.1
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.29
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.26.3
wcwidth==0.2.6
Werkzeug==2.3.6
xlrd==2.0.1
